refactor: 공통 클래스 패키징
