+++ src/main/java/maeilmail/question/QuestionApi.java
@@ -1,7 +1,7 @@
 package maeilmail.question;
 
 import lombok.RequiredArgsConstructor;
-import maeilmail.PaginationResponse;
+import maeilmail.support.PaginationResponse;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import org.springframework.data.web.PageableDefault;
