+++ src/main/java/maeilmail/subscribe/TemporalSubscribe.java
@@ -9,7 +9,7 @@ import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import maeilmail.BaseEntity;
+import maeilmail.support.BaseEntity;
 
 @Entity
 @Getter
