+++ src/main/java/maeilmail/subscribequestion/SubscribeQuestionApi.java
@@ -1,7 +1,7 @@
 package maeilmail.subscribequestion;
 
 import lombok.RequiredArgsConstructor;
-import maeilmail.PaginationResponse;
+import maeilmail.support.PaginationResponse;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.web.PageableDefault;
 import org.springframework.http.ResponseEntity;
