+++ src/main/java/maeilmail/subscribequestion/SubscribeQuestionQueryService.java
@@ -16,9 +16,9 @@ import com.querydsl.core.types.dsl.PathBuilder;
 import com.querydsl.jpa.impl.JPAQuery;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import lombok.RequiredArgsConstructor;
-import maeilmail.PaginationResponse;
 import maeilmail.question.Question;
 import maeilmail.question.QuestionCategory;
+import maeilmail.support.PaginationResponse;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
