+++ src/main/java/maeilmail/support/CacheConfig.java
@@ -1,4 +1,4 @@
-package maeilmail;
+package maeilmail.support;
 
 import java.util.Arrays;
 import java.util.List;
