+++ src/main/java/maeilmail/support/DistributedSupport.java
@@ -1,4 +1,4 @@
-package maeilmail;
+package maeilmail.support;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
