+++ src/main/java/maeilmail/support/JpaAuditingConfig.java
@@ -1,4 +1,4 @@
-package maeilmail;
+package maeilmail.support;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
