+++ src/main/java/maeilmail/support/PaginationResponse.java
@@ -1,4 +1,4 @@
-package maeilmail;
+package maeilmail.support;
 
 import java.util.List;
