+++ src/main/java/maeilmail/support/QueryDslConfig.java
@@ -1,4 +1,4 @@
-package maeilmail;
+package maeilmail.support;
 
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import jakarta.persistence.EntityManager;
