+++ src/test/java/maeilmail/question/QuestionQueryServiceTest.java
@@ -7,7 +7,7 @@ import static org.junit.jupiter.api.Assertions.assertAll;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.NoSuchElementException;
-import maeilmail.PaginationResponse;
+import maeilmail.support.PaginationResponse;
 import maeilmail.support.IntegrationTestSupport;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
