+++ src/test/java/maeilmail/subscribequestion/SubscribeQuestionQueryServiceTest.java
@@ -4,7 +4,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
 import java.util.List;
-import maeilmail.PaginationResponse;
+import maeilmail.support.PaginationResponse;
 import maeilmail.question.Question;
 import maeilmail.question.QuestionCategory;
 import maeilmail.question.QuestionRepository;
